LayoutTests/imported/w3c/web-platform-tests/fetch/api/response/response-consume-expected.txt


PASS Consume response's body: from text to text 
PASS Consume response's body: from text to blob 
PASS Consume response's body: from text to arrayBuffer 
PASS Consume response's body: from text to json 
FAIL Consume response's body: from text with correct multipart type to formData promise_test: Unhandled rejection with value: "Not implemented"
FAIL Consume response's body: from text without correct multipart type to formData (error case) assert_throws: function "function () { throw e }" threw "Not implemented" (undefined) expected object "TypeError" ("TypeError")
FAIL Consume response's body: from text with correct urlencoded type to formData promise_test: Unhandled rejection with value: "Not implemented"
FAIL Consume response's body: from text without correct urlencoded type to formData (error case) assert_throws: function "function () { throw e }" threw "Not implemented" (undefined) expected object "TypeError" ("TypeError")
PASS Consume response's body: from blob to blob 
PASS Consume response's body: from blob to text 
PASS Consume response's body: from blob to arrayBuffer 
PASS Consume response's body: from blob to json 
FAIL Consume response's body: from blob with correct multipart type to formData promise_test: Unhandled rejection with value: "Not implemented"
FAIL Consume response's body: from blob without correct multipart type to formData (error case) assert_throws: function "function () { throw e }" threw "Not implemented" (undefined) expected object "TypeError" ("TypeError")
FAIL Consume response's body: from blob with correct urlencoded type to formData promise_test: Unhandled rejection with value: "Not implemented"
FAIL Consume response's body: from blob without correct urlencoded type to formData (error case) assert_throws: function "function () { throw e }" threw "Not implemented" (undefined) expected object "TypeError" ("TypeError")
FAIL Consume response's body: from FormData to formData promise_test: Unhandled rejection with value: "Not implemented"
FAIL Consume response's body: from FormData without correct type to formData (error case) assert_throws: function "function () { throw e }" threw "Not implemented" (undefined) expected object "TypeError" ("TypeError")
FAIL Consume response's body: from FormData to blob promise_test: Unhandled rejection with value: undefined
FAIL Consume response's body: from FormData to text promise_test: Unhandled rejection with value: undefined
FAIL Consume response's body: from FormData to arrayBuffer promise_test: Unhandled rejection with value: undefined
FAIL Consume response's body: from URLSearchParams to formData promise_test: Unhandled rejection with value: "Not implemented"
FAIL Consume response's body: from URLSearchParams without correct type to formData (error case) assert_throws: function "function () { throw e }" threw "Not implemented" (undefined) expected object "TypeError" ("TypeError")
FAIL Consume response's body: from URLSearchParams to blob assert_equals: Blob body type should be computed from the response Content-Type expected "application/x-www-form-urlencoded;charset=utf-8" but got "application/x-www-form-urlencoded"
PASS Consume response's body: from URLSearchParams to text 
PASS Consume response's body: from URLSearchParams to arrayBuffer 
PASS Consume response's body: from stream to blob 
PASS Consume response's body: from stream to text 
PASS Consume response's body: from stream to arrayBuffer 
PASS Consume response's body: from stream to json 
FAIL Consume response's body: from stream with correct multipart type to formData promise_test: Unhandled rejection with value: "Not implemented"
FAIL Consume response's body: from stream without correct multipart type to formData (error case) assert_throws: function "function () { throw e }" threw "Not implemented" (undefined) expected object "TypeError" ("TypeError")
FAIL Consume response's body: from stream with correct urlencoded type to formData promise_test: Unhandled rejection with value: "Not implemented"
FAIL Consume response's body: from stream without correct urlencoded type to formData (error case) assert_throws: function "function () { throw e }" threw "Not implemented" (undefined) expected object "TypeError" ("TypeError")
PASS Consume response's body: from fetch to blob 
PASS Consume response's body: from fetch to text 
PASS Consume response's body: from fetch to arrayBuffer 
FAIL Consume response's body: from fetch without correct type to formData (error case) assert_throws: function "function () { throw e }" threw "Not implemented" (undefined) expected object "TypeError" ("TypeError")
FAIL Consume response's body: from multipart form data blob to formData promise_test: Unhandled rejection with value: "Not implemented"
